make category a required field for transactions and protect all api routes with authentication

+++ app/api/accounts/[id]/route.ts
@@ -1,17 +1,27 @@
 import { prisma } from '@/lib/prisma';
 import { NextResponse } from 'next/server';
+import { getServerSession } from "next-auth";
+import { requireAuthentication } from '@/lib/requireAuthentication';
 
-//TODO: add authentication and authorization to ensure users can only access their own accounts
 // update one specific existing account by id
 
 export async function PUT(request: Request,
     { params }: { params: Promise<{ id: string }>  }
   ) {
+    const user = await requireAuthentication();
+    if (!user) return new Response("Unauthorized", { status: 401 });
     const data = await request.json();
     console.log("Received data for account update:");
     try{
         const { id } = await params;
-        const account = await prisma.account.update({
+        const account = await prisma.account.findUnique({
+            where: {id: id },
+        });
+
+        if(!account || account.userId !== user.id){
+            return new Response("Unauthorized", { status: 401 });
+        }
+        const updatedAccount = await prisma.account.update({
         where: { id: id },
         data: {
             ...(data.name && { name: data.name }),
@@ -19,8 +29,8 @@ export async function PUT(request: Request,
             ...(data.balance !== undefined && { balance: data.balance }),
         }
         });
-        console.log("Updated account:", account);
-        return NextResponse.json({message: 'Account updated successfully', data: account}, {status: 200});
+        console.log("Updated account:", updatedAccount);
+        return NextResponse.json({message: 'Account updated successfully', data: updatedAccount}, {status: 200});
     } catch (error) {
         console.log("Error updating account:", error);
         return NextResponse.json({message: 'Error updating account, ${error.message}'}, {status: 500});
@@ -32,7 +42,16 @@ export async function DELETE(
     req: Request,
     {params}: {params: Promise<{id: string}>}
 ){
+    const user = await requireAuthentication();
+    if (!user) return new Response("Unauthorized", { status: 401 });
+    const { id } = await params;
+    const account = await prisma.account.findUnique({
+        where: {id: id },
+    });
 
+    if(!account || account.userId !== user.id){
+        return new Response("Unauthorized", { status: 401 });
+    }
     try{
         const { id } = await params;
         await prisma.account.delete({
