+++ app/api/categories/[id]/route.ts
@@ -1,5 +1,7 @@
+import { getServerSession } from "next-auth";
 import { prisma } from '@/lib/prisma';
 import { NextResponse } from 'next/server';
+import { requireAuthentication } from "@/lib/requireAuthentication";
 
 //TODO: add authentication and authorization to ensure users can only access their own accounts
 // update one specific existing account by id
@@ -7,19 +9,29 @@ import { NextResponse } from 'next/server';
 export async function PUT(request: Request,
     { params }: { params: Promise<{ id: string }>  }
   ) {
+    
     const data = await request.json();
-    console.log("Received data for account update:");
+    const user = await requireAuthentication();
+    if (!user) return new Response("Unauthorized", { status: 401 });
+    console.log("Received data for category update:");
     try{
         const { id } = await params;
-        const category = await prisma.category.update({
+        const category = await prisma.category.findUnique({
+            where: { id: id},
+        })
+        if(!category || category.userId !== user.id){
+            return new Response("Unauthorized", { status: 401 });
+        }
+
+        const updatedCategory = await prisma.category.update({
         where: { id: id },
         data: {
             ...(data.name && { name: data.name }),
             ...(data.type && { type: data.type }),
         }
         });
-        console.log("Updated category:", category);
-        return NextResponse.json({message: 'Category updated successfully', data: category}, {status: 200});
+        console.log("Updated category:", updatedCategory);
+        return NextResponse.json({message: 'Category updated successfully', data: updatedCategory}, {status: 200});
     } catch (error) {
         return NextResponse.json({message: 'Error updating category, ${error.message}'}, {status: 500});
     }
@@ -30,9 +42,16 @@ export async function DELETE(
     req: Request,
     {params}: {params: Promise<{id: string}>}
 ){
-
+    const user = await requireAuthentication();
+    if (!user) return new Response("Unauthorized", { status: 401 });
     try{
         const { id } = await params;
+        const category = await prisma.category.findUnique({
+            where: { id: id},
+        })
+        if(!category || category.userId !== user.id){
+            return new Response("Unauthorized", { status: 401 });
+        }
         await prisma.category.delete({
         where: {id: id},
     });
