+++ app/api/transactions/[id]/route.ts
@@ -1,5 +1,8 @@
+import { getServerSession } from "next-auth";
 import { prisma } from '@/lib/prisma';
 import { NextResponse } from 'next/server';
+import { requireAuthentication } from "@/lib/requireAuthentication";
+import { trace } from "console";
 
 //TODO: add authentication and authorization to ensure users can only access their own accounts
 // update one specific existing account by id
@@ -8,10 +11,32 @@ export async function PUT(request: Request,
     { params }: { params: Promise<{ id: string }>  }
   ) {
     const data = await request.json();
-    console.log("Received data for account update:");
+    const user = await requireAuthentication();
+    if (!user) return new Response("Unauthorized", { status: 401 });
+    const account = await prisma.account.findUnique({
+        where: { id: data.accountId },
+      });
+      
+    const category = await prisma.category.findUnique({
+        where: { id: data.categoryId },
+    });
+    if(!category || category.userId !== user.id){
+        return new Response("Invalid category", {status: 400})
+    }
+    if (!account || account.userId !== user.id) {
+        return new Response("Invalid account", { status: 400 });
+    }
     try{
         const { id } = await params;
-        const account = await prisma.transaction.update({
+        const transaction = await prisma.transaction.findUnique({
+            where: { id: id }
+        })
+
+        if(!category || category.userId !== user.id){
+            return new Response("Unauthorized", { status: 401 });
+        }
+
+        const updatedTransaction = await prisma.transaction.update({
         where: { id: id },
         data: {
             ...(data.amount && { amount: data.amount }),
@@ -21,8 +46,8 @@ export async function PUT(request: Request,
             ...(data.categoryId && { categoryId: data.categoryId }),
         }
         });
-        console.log("Updated transaction:", account);
-        return NextResponse.json({message: 'Transaction updated successfully', data: account}, {status: 200});
+        console.log("Updated transaction:", updatedTransaction);
+        return NextResponse.json({message: 'Transaction updated successfully', data: updatedTransaction}, {status: 200});
     } catch (error) {
         return NextResponse.json({message: 'Error updating transaction, ${error.message}'}, {status: 500});
     }
@@ -33,9 +58,19 @@ export async function DELETE(
     req: Request,
     {params}: {params: Promise<{id: string}>}
 ){
-
+    const user = await requireAuthentication();
+    if (!user) return new Response("Unauthorized", { status: 401 });
+    
     try{
         const { id } = await params;
+        const transaction = await prisma.transaction.findUnique({
+            where: { id: id },
+            include: { account: true },
+        });
+        
+        if (!transaction || transaction.account.userId !== user.id){
+            return new Response("Unauthorized", { status: 401 });
+        }
         await prisma.transaction.delete({
         where: {id: id},
     });
